WebIDL/tests/submissions/heycam/unrestricted-double-004.html

<!DOCTYPE html>
<title>unrestricted double</title>
<script src="http://w3c-test.org/resources/testharness.js"></script>
<script src="http://w3c-test.org/resources/testharnessreport.js"></script>
<script src="utils.js"></script>
<div id="log"></div>

<script>
// 4.2.15 unrestricted double
//
// IDL unrestricted double values are represented with JS number values.

// feature: HTMLInputElement.number (attribute) (of type unrestricted double)

test(function() {
  var e = document.createElement("input");
  e.type = "number";
  assert_equals(typeof e.valueAsNumber, "number");
}, "HTMLInputElement.valueAsNumber returns a value of the correct type");

// feature: Float64Array (indexed property getter) (of type unrestricted double)

test(function() {
  var a = new Float64Array(1);
  assert_equals(typeof a[0], "number");
}, "Float64Array indexed getter returns a value of the correct type");
</script>
